refactor: move files insertion logic to hook
